Correctly refactor unit tests
The unit test refactoring left copy/pasted code.  This commit splits
the common piece into a single header, allowign `s3_tests.cc` with
the unit tests that utilize AWS S3 while `s3_unit_tests.cc` use the
minio instance started up by ctest.
